back-end/requirements.txt
# Core dependencies
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-qdrant==0.2.0
openai==1.75.0

# Vector databases
qdrant-client==1.14.2

# Data processing
numpy==2.2.4
pydantic==2.11.3


# Web framework (if you're using FastAPI)
fastapi==0.115.9
uvicorn==0.34.1

# Utilities
python-dotenv==1.1.0


